requirements.txt
Django==3.2.18
face-recognition==1.3.0
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.5.0